requirements.txt
flask==2.3.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.0
pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0 